+++ app/controllers/project_settings_controller.rb
@@ -8,6 +8,11 @@ class ProjectSettingsController < ApplicationController
   before_action :set_settings_project
 
   def show
+    if legacy_archive_search_path?
+      redirect_to archives_project_path(@project, legacy_archive_search_params)
+      return
+    end
+
     @settings_sections = settings_sections_for_current_user
     @settings_section = normalized_settings_section
     load_project_settings_context
@@ -15,6 +20,21 @@ class ProjectSettingsController < ApplicationController
   end
 
   private
+
+  def legacy_archive_search_path?
+    @project.managed_by?(current_user) &&
+      params[:section] == "data" &&
+      params[:archive_path].in?(%w[search_archives investigations])
+  end
+
+  def legacy_archive_search_params
+    raw_search = params[:archive_search]
+    return {} unless raw_search.respond_to?(:permit)
+
+    {
+      archive_search: raw_search.permit(*ProjectArchiveInvestigationSearch::SEARCH_FIELDS).to_h
+    }
+  end
 
   def set_settings_project
     @project = if admin_user?